Change pdf preview follow REST convention
