--- app/views/pdf_templates/sites/template.html.haml
@@ -1,1 +1,0 @@
-!= @pdf_template.render(@site)
