+++ app/views/sites/pdf_templates/show.html.haml
@@ -1,0 +1,1 @@
+!= @pdf_template.render(@site)
